requirements.txt
paho-mqtt
opencv-python
Pillow
